Separate graphql queries and mutations gql tags from js code
