+++ client/src/components/mainPage/MainPage.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import { useQuery } from '@apollo/react-hooks'
-import gql from 'graphql-tag'
 import { makeStyles } from '@material-ui/styles'
 import { Card, CardContent, Typography } from '@material-ui/core'
  
+import GET_CATEGORIES from '../../queries/GetCategories'
+
 import Category from './Category'
 import Loading from '../Loading'
 
@@ -12,28 +13,6 @@ const useStyles = makeStyles({
     marginTop: 15
   }
 })
-
-const GET_CATEGORIES = gql`
-  {
-    categories {
-      id
-      name
-      topics {
-        id
-        name
-        createdAt
-        user {
-          id
-          name
-        }
-      }
-      user {
-        id
-        name
-      }
-    }
-  }
-`
 
 const MainPage = () => {
   const classes = useStyles()
