+++ client/src/queries/GetPost.js
@@ -1,0 +1,28 @@
+import gql from 'graphql-tag'
+
+const GET_POST = gql`
+  query getPost($id: ID!) {
+    post(id: $id) {
+      id
+      title
+      content
+      createdAt
+      comments {
+        id
+        content
+        createdAt
+        user {
+          id
+          name
+        }
+      }
+      user {
+        id
+        name
+        role
+      }
+    }
+  }
+`
+
+export default GET_POST
