+++ client/src/queries/GetTopicNamesAndIds.js
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag'
+
+const GET_TOPIC_NAMES_AND_IDS = gql`
+  {
+    topics {
+      id
+      name
+    }
+  }
+`
+
+export default GET_TOPIC_NAMES_AND_IDS
